docs/20260330_patch/QCVIZ_MOLCHAT_CORE_BUNDLE_2026-03-30/QCViz/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qcviz-mcp"
version = "0.2.0"
description = "Quantum Chemistry Visualization MCP server with FastAPI, PySCF, 3Dmol.js, WebSocket, and LLM planning"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "QCViz Team" }
]
dependencies = [
  "fastapi>=0.110,<1.0",
  "uvicorn[standard]>=0.29,<1.0",
  "jinja2>=3.1,<4.0",
  "pydantic>=2.6,<3.0",
  "numpy>=1.26",
  "scipy>=1.11",
  "pyscf>=2.4,<3.0",
  "python-dotenv>=1.0,<2.0",
  "httpx>=0.27,<1.0",
  "orjson>=3.10,<4.0",
]

[project.optional-dependencies]
llm-openai = [
  "openai>=1.30,<2.0",
]
llm-gemini = [
  "google-genai>=0.7,<2.0",
]
worker = [
  "redis>=5.0,<6.0",
  "arq>=0.26,<1.0",
]
dev = [
  "pytest>=8.0,<9.0",
  "pytest-asyncio>=0.23,<1.0",
  "pytest-cov>=5.0,<6.0",
  "wsproto>=1.2,<2.0",
  "httpx-ws>=0.7,<1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "slow: PySCF 실제 계산 포함 (느림)",
    "integration: 통합 테스트",
    "e2e: End-to-End 테스트",
    "live: 실제 외부 API 스모크 테스트",
]
timeout = 60

[tool.coverage.run]
source = ["src/qcviz_mcp"]
branch = true
